Trade: rework and extend PhongMaterialData.
It now supports both color+texture (instead of just one or the other)
which made it much simpler to use (and implement as well), also can
reference normal maps and specify texture coordinate transformation now.

The old APIs are depecated, but no tests or other code is updated yet in
order to ensure everything continues to work as before.