Shaders: rework Phong to support directional and attenuated point lights.
This is a -- long overdue -- breaking change to the rendering output of
this shader, finally adding support for lights that get darker over
distance. The attenuation equation is basically what's documented in
LightData, and the distinction between directional and point lights is
made using a newly added the fourth component of position (which means
the old three-component setters are all deprecated). This allows the
shader code to be practically branchless, which I find to be nice.

This breaks basically all rendering output so all existing Phong and
MeshTools::compile() test outputs had to be regenerated.