Shaders: fix distance attentuation calculation in Phong.
Not that either way would be more correct than the other (this is what
three.js uses I think), but it was documented everywhere to be

    1/(1 + d^2)

but the calculation instead did

   1/(1 + d)^2

This now also means the analytical test equation works. I should have
paid more attention to it not matching before, because that obviously
pointed to this problem.

Point lights are now significantly brighter than before, the tests were
updated to use a larger distance to avoid issues with overflows. Does
not affect the (default) directional lights in any way.