Vk: rename PipelineStage::AllGraphics to AllRasterization.
Because that's what it is, after all. Also mention this naming
difference in the Vulkan wrapping overview.

I was wondering if I should rename QueueFlag::Graphics too, but since
graphics queues are used for (not really rasterization-specific) image
copies as well, and ray tracing is actually done on compute queues, I
don't think renaming to Rasterization is a good idea. Clarified the
comments at least.