Platform: fix a crash in GlfwApplication::setCursor() on GLFW 3.4.
There's four more new cursors and the _cursors array was too small.

At first I got confused because I thought the assertion on top is done
against the CursorMap, which didn't contain the Hidden cursors. So to
avoid confusing myself again in the future, I moved the assert after the
special cases and made both arrays the same size since it doesn't make
sense to have always-empty fields in there.

Similar change is done in Sdl2Application, and an assert is added to
avoid a nondescript crash if the window is not created yet.