Trade: reverse AbstractSceneConverter::convertToFile() argument order.
It should be input first, output second, like with all other APIs. I
remember I was trying something else here, but that didn't really make
sense in the end. Also took that opportunity to get rid of one
std::string.

The original signature is a deprecated alias to the new one and will be
removed in a future release.