Propagate configuration options in Any* plugins.
Minor but very important convenience feature, especially useful when
dealing with command-line apps. This now works:

   magnum-imageconverter a.png a.jpg -c jpegQuality=0.75

The AnyImageConverter gets the jpegQuality option and then
automatically propagates it to the concrete plugin (which is either
JpegImageConverter or StbImageConverter), possibly warning in case the
target plugin doesn't recognize given option (i.e., doesn't list it in
its default configuration). Previously the user had to always specify a
concrete converter implementation using -C, which was rather annoying
and nonintuitive.