GL: the multi-draw test for Mesh was underwhelming, to say the least.
Or, the horror when I comment-out half of it and it still kinda passes.
The new code tests it more thoroughly, with actually checking that
different draw produce different results instead of one draw overwriting
the other.

Plus also verifying interaction with gl_VertexID and gl_DrawID to ensure
it really behaves correctly:

 - gl_VertexID should get the same value regardless of whether we
   perform a multi-draw or a one-by-one fallback, so this catches issues
   where the mesh would get reconfigured in some nasty way, losing this
   information
 - gl_DrawID should get properly populated if and only if we use the
   multi-draw code path, so this catches issues where the fallback would
   get used instead of the actual multi-draw code path