Audio,Text,Trade: accept void ArrayViews in importer plugins.
This makes it much less annoying to pass arbitrarily typed data, such as
std::uint8_t or char8_t and what not. It was already done like this for
the new shader converter plugins, where the input is often 32-bit ints
for SPIR-V.

OTOH the internal virtual API is kept with ArrayView<const char>, as
that makes it easier to operate on by the implementations.