Trade: remove the useless indirection from SceneField::Parent.
This got originally added as some sort of a kludge to make it easy to go
to the parent transformation, assuming Parent and Transformation share
the same object mapping:

    parentTransformation = transformations[parents[i]]

But after some ACTUAL REAL WORLD use, I realized that there's often a
set of objects that have a Parent defined, and then another, completely
disjoint, set of objects that have a transformation (for example certain
nodes having no transformation at all because it's an identity). And so
this parent indirection is not only useless, but in fact an additional
complication. Let's say we make a map of the transformations, where
transformationMap[i] is a transformation for object i:

    transformationMap = {}
    for j in range(len(transformations)):
        transformationMap[transformationObject[j]] = transformation[j]

Then, with *no* assumptions about shared object mapping, the indirection
would cause parent transformation retrieval to look like this:

    parentTransformation = transformationMap[parentObjects[parents[i]]

While *without* the indirection, it'd be just

    parentTransformation = transformationMap[parents[i]]