GL: fix EXT_color_buffer_half_float to be presennt also in WebGL 2.
It was mistakenly thought to be replaced by the EXT_color_buffer_float
(which replaces WEBGL_color_buffer_float and in addition lists both 16-
and 32-bit float variants). But since there are still those stupid
patents for rendering to 32-bit float attachments, certain hardware
supports only rendering to 16-bit and not 32-bit, so the "superset"
extension isn't enough to be able to discover which hardware can
render to half-floats.

Also updated (hopefully all) docs to list this extension as being an
option on WebGL 2 as well.