Add an ability to store implementation-specific values in MeshIndexType.
The type is now extended to 32 bits. In the GL and Vk libraries it means
one can now do things like

    MeshIndexType type = meshIndexTypeWrap(GL_UNSIGNED_BYTE);

and passing that to GL::Mesh or Vk::Mesh will cause it to use the value
directly, instead of doing a mapping from a generic type. The *real* use
case for this is however to allow custom index buffer representations in
Trade::MeshData. Support for that will be hooked up in the following
commit.