Shaders: implement object ID texture visualization in MeshVisualizer.
Hah this took a while, as there was no texture scaffolding in place at
all. Thus all this had to be added and tested for the first time:

 * 2D textures
 * 2D texture arrays
 * Texture transformation uniforms
 * Texture transformation UBOs
 * Instanced texture offset

This also means that MeshVisualizer can be used to visualize arbitrary
(single-channel) integer textures now, not just render meshes with
object ID textures. Yay for feature parity!